11 спринт второе ревью
