+++ src/main/java/ru/yandex/practicum/filmorate/storage/user/InMemoryUserStorage.java
@@ -7,11 +7,11 @@ import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
 import ru.yandex.practicum.filmorate.exceptions.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.time.LocalDate;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -32,14 +32,6 @@ public class InMemoryUserStorage implements UserStorage {
     public User create(User user) {
         log.info("Обработка Create-запроса...");
         duplicateCheck(user);
-
-        validateEmail(user.getEmail());
-        validateLogin(user.getLogin());
-        if (user.getName() == null || user.getName().isBlank()) {
-            user.setName(user.getLogin());
-        }
-        validateBirthday(user.getBirthday());
-
         user.setId(getNextId());
         user.setFriends(new HashSet<>());
         users.put(user.getId(), user);
@@ -48,10 +40,6 @@ public class InMemoryUserStorage implements UserStorage {
 
     @Override
     public User update(User newUser) {
-        if (newUser.getId() == null) {
-            throw new ValidationException("Id должен присутствовать");
-        }
-
         if (!users.containsKey(newUser.getId())) {
             throw new NotFoundException("Пользователь с id = " + newUser.getId() + " не найден");
         }
@@ -81,28 +69,6 @@ public class InMemoryUserStorage implements UserStorage {
                 throw new DuplicatedDataException("Эта почта уже используется");
             }
         }
-    }
-
-    private void validateEmail(String email) {
-        if (email == null || email.isBlank() || !email.contains("@") || email.contains(" ") || email.length() == 1) {
-            throw new ValidationException("Неверный формат электронной почты");
-        }
-    }
-
-    private void validateLogin(String login) {
-        if (login == null || login.contains(" ") || login.isBlank()) {
-            throw new ValidationException("Логин не должен быть пустым или состоять из пробелов");
-        }
-    }
-
-    private void validateBirthday(LocalDate birthday) {
-        if (birthday == null) {
-            throw new ValidationException("Дата рождения должна быть указана");
-        }
-        if (birthday.isAfter(LocalDate.now())) {
-            throw new ValidationException("Дата рождения не должна быть из будущего");
-        }
-
     }
 
     @Override
@@ -159,7 +125,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public Collection<User> getFriends(Long id) throws NotFoundException {
+    public List<User> getFriends(Long id) throws NotFoundException {
         User user = findById(id);
 
         if (user.getFriends() == null || user.getFriends().isEmpty()) {
@@ -172,7 +138,7 @@ public class InMemoryUserStorage implements UserStorage {
     }
 
     @Override
-    public Collection<User> getCommonFriends(Long userId, Long otherUserId) throws NotFoundException {
+    public List<User> getCommonFriends(Long userId, Long otherUserId) throws NotFoundException {
         User user = findById(userId);
         User otherUser = findById(otherUserId);
